API - Add AutoMapper, FluentValidation e DI
API - Adicionando AutoMapper, FluentValidation, injeção de dependência dos projetos 'Infrastructure' e 'API' e refatoração do Sqlite.

+++ API/FilmesApp/FilmesApp.API/Controllers/MoviesController.cs
@@ -1,8 +1,12 @@
-﻿using FilmesApp.Core.Entities;
+﻿using AutoMapper;
+using FilmesApp.API.DTOs;
+using FilmesApp.Core.Entities;
 using FilmesApp.Core.Interfaces.Services;
+using FluentValidation;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace FilmesApp.API.Controllers
@@ -12,10 +16,14 @@ namespace FilmesApp.API.Controllers
     public class MoviesController : ControllerBase
     {
         private IMovieService _movieService;
+        private IMapper _mapper;
+        private IValidator<SaveMovieDTO> _validator;
 
-        public MoviesController(IMovieService movieService)
+        public MoviesController(IMovieService movieService, IMapper mapper, IValidator<SaveMovieDTO> validator)
         {
             _movieService = movieService;
+            _mapper = mapper;
+            _validator = validator;
         }
 
         [HttpGet("{id}")]
@@ -31,7 +39,8 @@ namespace FilmesApp.API.Controllers
                 if (movie == null)
                     return NotFound();
 
-                return Ok(movie);
+                var movieDTO = _mapper.Map<Movie, MovieDTO>(movie);
+                return Ok(movieDTO);
             }
             catch (Exception ex)
             {
@@ -48,7 +57,9 @@ namespace FilmesApp.API.Controllers
             {
                 var movies = await _movieService.GetAll();
 
-                return Ok(movies);
+                var moviesDTO = _mapper.Map<IEnumerable<Movie>, IEnumerable<MovieDTO>>(movies);
+
+                return Ok(moviesDTO);
             }
             catch (Exception ex)
             {
@@ -60,10 +71,16 @@ namespace FilmesApp.API.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult> Add([FromBody] Movie movie)
+        public async Task<ActionResult> Add([FromBody] SaveMovieDTO saveMovie)
         {
             try
             {
+                var validatorResult = await _validator.ValidateAsync(saveMovie);
+
+                if (!validatorResult.IsValid)
+                    return BadRequest(validatorResult.Errors);
+
+                var movie = _mapper.Map<SaveMovieDTO, Movie>(saveMovie);
                 await _movieService.Save(movie);
 
                 return Ok();
@@ -74,7 +91,7 @@ namespace FilmesApp.API.Controllers
             }
         }
 
-        [HttpDelete]
+        [HttpDelete("{id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
@@ -82,6 +99,11 @@ namespace FilmesApp.API.Controllers
         {
             try
             {
+                var movie = await _movieService.GetById(id);
+
+                if (movie == null)
+                    return NotFound();
+
                 await _movieService.Delete(id);
 
                 return Ok();
